feat(security): wire TerraFusion.Security module into solution + harden interfaces
- Add TerraFusion.Security.csproj to solution with project reference from API
- Expand IAuditService, IAuthenticationService, IMfaService, ISessionManager
  with NIST 800-63B compliant method signatures
- Add AuditModels (AuditEntry, SecurityEvent, ComplianceReport types)
- Add Session, UserInfo, IHashingService domain models
- Implement ProductionAuditService with structured logging
- Add SecurityStubs for development/test environments
- Add GlobalUsings for consistent namespace imports
- Wire SecurityExtensions into DI registration (Program.cs)
- Normalize PasswordModels for lockout + history tracking
- Fix SecurityController and SecurityMiddleware compilation

23 files, 529 insertions — completes Sprint 1 security foundation.

+++ backend/TerraFusion.Security/Controllers/SecurityController.cs
@@ -112,7 +112,7 @@ public class SecurityController : ControllerBase
             var response = new SecurityEventsResponse
             {
                 Events = events,
-                TotalCount = totalCount,
+                TotalCount = (int)totalCount,
                 Page = page,
                 PageSize = pageSize,
                 TotalPages = (int)Math.Ceiling((double)totalCount / pageSize)
